+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+* Fixed zoom not following cursor if the map wasn't moved.
 * Fixed spurious "Error from IO runtime" at shutdown.
 
 ## 0.29.0