+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
 * Add options for double click to zoom in and out
+* Fixed zoom not following cursor if the map wasn't moved.
+* Fixed spurious "Error from IO runtime" at shutdown.
+
 
 ## 0.29.0