+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
 * Add option to zoom without holding ctrl on native and web
+* Fixed zoom not following cursor if the map wasn't moved.
+* Fixed spurious "Error from IO runtime" at shutdown.
 
 ## 0.29.0