+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+* Add option to zoom without holding ctrl on native and web
 * Fixed zoom not following cursor if the map wasn't moved.
 * Fixed spurious "Error from IO runtime" at shutdown.